requirements.txt
# requirements.txt
# pip install -r requirements.txt
argparse>=1.4.0
Pillow>=3.3.0
